requirements.txt
# Web服务依赖
Flask==3.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
eventlet==0.33.3
gunicorn==21.2.0
httpx==0.24.1

# 交易所与AI依赖
# ccxt 已移除，改用 OKXClient（直接调用 OKX API）
openai==1.6.0

# 数据处理依赖
pandas>=2.2.3  # Python 3.13 兼容版本
numpy

# 工具依赖
python-dotenv==1.0.0
requests==2.31.0
urllib3
